nexus/pyproject.toml
[project]
name = "khala"
version = "0.1.0"
description = "Enterprise RAG + GraphRAG - grounded knowledge retrieval with Korean NLP + OTel integration"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # CLI
    "typer[all]>=0.12.0",
    # Database
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    # Korean NLP
    "mecab-python3>=1.0.9",
    # HTTP client (Ollama, Tempo)
    "httpx>=0.27.0",
    # LLM
    "anthropic>=0.39.0",
    # Utilities
    "pyyaml>=6.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "tiktoken>=0.8.0",
    "python-frontmatter>=1.1.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
khala = "khala.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
